Using proper TMDB queries and fixed adapter updating
Changed to using separate TMDB queries for the different modes instead of just sorting a DISCOVER query. Fixed the name of TMDB from TMBD and TVBD in certain cases. THe new query method also required properly updating the adapter instead of just making a new one on the HTTP response.